remove cloudshell from redux and now inject value
