+++ src/js/components/cloud-shell/cloud-shell.js
@@ -1,0 +1,207 @@
+import React, { useState, useEffect } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import { Resizable } from 're-resizable';
+import IconButton from '@material-ui/core/IconButton';
+import CloseIcon from '@material-ui/icons/Close';
+import Autorenew from '@material-ui/icons/Autorenew';
+import KeyboardIcon from '@material-ui/icons/Keyboard';
+import LoopSharpIcon from '@material-ui/icons/LoopSharp';
+import OpenInNewIcon from '@material-ui/icons/OpenInNew';
+import DeleteIcon from '@material-ui/icons/Delete';
+import './cloud-shell.scss';
+import { withStyles } from '@material-ui/core';
+import { axiosAuth } from 'js/utils';
+import api from 'js/redux/api';
+import { creerNouveauService, requestDeleteMonService } from 'js/redux/actions';
+import { getMinioToken } from 'js/minio-client';
+import { getVaultToken } from 'js/vault-client';
+import {
+	getOptions,
+	getValuesObject,
+} from 'js/components/my-lab/catalogue/catalogue-navigation/leaf/deploiement/nouveau-service';
+
+const CloudShell = () => {
+	const user = useSelector((store) => store.user);
+	const [cloudShellStatus, setCloudShellStatus] = useState();
+	const [url, setUrl] = useState();
+	const [height, setHeight] = useState(200);
+	const [visibility, setVisibility] = useState(false);
+	const [minioCredentials, setMinioCredentials] = useState();
+	const dispatch = useDispatch();
+	const intervalId = React.useRef(0);
+	var deleting = false;
+
+	// if (status === 'DOWN' && deleting === true) {
+	// 	clearInterval(intervalId.current);
+	// 	deleting = false;
+	// } else if (status === 'UP' && deleting === false) {
+	// 	clearInterval(intervalId.current);
+	// 	intervalId.current = null;
+	// }
+
+	const launchCloudShell = (user) => {
+		axiosAuth(`${api.cloudShell}`).then((data) => {
+			const service = data.packageToDeploy;
+			setCloudShellStatus(data.status);
+			if (data.status === 'DOWN') {
+				dispatch(
+					creerNouveauService(
+						{
+							...service,
+							catalogId: data.catalogId,
+						},
+						getValuesObject(getOptions(user, service, minioCredentials, {}).fV),
+						false
+					)
+				).then((response) => {
+					setUrl(response[0].labels.ONYXIA_URL);
+					setCloudShellStatus('UP');
+				});
+			} else {
+				if (data.status === 'UP') {
+					setUrl(data.url);
+				}
+			}
+		});
+		return launchCloudShell;
+	};
+
+	const deleteCloudShell = (idep) => {
+		dispatch(requestDeleteMonService({ id: 'cloudshell' }));
+		setCloudShellStatus();
+		setUrl();
+	};
+
+	useEffect(() => {
+		if (!minioCredentials) {
+			getMinioToken()
+				.then((credentials) => {
+					setMinioCredentials(credentials);
+				})
+				.catch((e) => {
+					setMinioCredentials({});
+				});
+		}
+	}, [minioCredentials]);
+
+	if (!visibility || cloudShellStatus === 'DOWN') {
+		const button =
+			cloudShellStatus && cloudShellStatus === 'DOWN' ? (
+				<LoopSharpIcon className="loading" />
+			) : (
+				<KeyboardIcon />
+			);
+		return (
+			<div style={{ position: 'fixed', bottom: 0, zIndex: 999 }}>
+				<IconButton
+					aria-label="maximize"
+					onClick={() => {
+						if (
+							(!cloudShellStatus || cloudShellStatus === 'DOWN') &&
+							deleting === false
+						) {
+							intervalId.current = launchCloudShell(user);
+						}
+						setVisibility(true);
+					}}
+					className="maximize-shell"
+				>
+					{button}
+				</IconButton>
+			</div>
+		);
+	}
+
+	var content = null;
+	if (cloudShellStatus) {
+		content = (
+			<Resizable
+				size={{
+					height: { height },
+					width: '100%',
+				}}
+				onResizeStop={(e, direction, ref, d) => {
+					setHeight(height + d.height);
+				}}
+			>
+				<iframe
+					title="Cloud shell"
+					height={height}
+					width="100%"
+					src={url}
+					id="cloudshell-iframe"
+				></iframe>
+			</Resizable>
+		);
+	}
+
+	const CloudShellIconButton = withStyles({
+		root: {
+			color: 'white',
+		},
+	})(IconButton);
+
+	return (
+		<div style={{ position: 'fixed', bottom: 0, zIndex: 999, width: '100%' }}>
+			<div
+				style={{
+					width: 'fit-content',
+					borderTopRightRadius: '10px',
+					backgroundColor: 'rgba(0, 0, 0, 0.35)',
+				}}
+			>
+				<CloudShellIconButton
+					aria-label="autorenew"
+					onClick={() =>
+						document
+							.getElementById('cloudshell-iframe')
+							.parentNode.replaceChild(
+								document.getElementById('cloudshell-iframe').cloneNode(),
+								document.getElementById('cloudshell-iframe')
+							)
+					}
+					className="renew-shell"
+				>
+					<Autorenew />
+				</CloudShellIconButton>
+
+				<CloudShellIconButton
+					aria-label="openinnewicon"
+					onClick={() => {
+						window.open(
+							String(document.getElementById('cloudshell-iframe').src),
+							'_blank'
+						);
+						setVisibility(false);
+					}}
+					className="opennewtab-shell"
+				>
+					<OpenInNewIcon />
+				</CloudShellIconButton>
+
+				<CloudShellIconButton
+					aria-label="delete"
+					onClick={() => {
+						deleting = true;
+						setVisibility(false);
+						deleteCloudShell(user.IDEP);
+					}}
+					className="close-shell"
+				>
+					<DeleteIcon />
+				</CloudShellIconButton>
+
+				<CloudShellIconButton
+					aria-label="close"
+					onClick={() => setVisibility(false)}
+					className="close-shell"
+				>
+					<CloseIcon />
+				</CloudShellIconButton>
+			</div>
+			{content}
+		</div>
+	);
+};
+
+export default CloudShell;
