+++ src/js/redux/actions/index.js
@@ -3,5 +3,4 @@ export * from './my-lab';
 export * from './mes-fichiers';
 export * from './secrets';
 export * from './constantes';
-export * from './cloud-shell';
 export * from './s3';
